New: KeyGeometryUpdater
RevBy: Kimmo Surakka, Pekka Vuorela
Details: Moved geomtry handling outside of MImKeyArea, hoping it makes things
easier to follow. Realized that MImKey needs to fix its cached rect behaviour.
Added tests for that, too.

+++ m-keyboard/widgets/mimkeyarea.h
@@ -119,7 +119,7 @@ private:
     KeyRowList rowList; //!< stores all rows of this key area
     qreal cachedWidgetHeight; //!< cached widget height
     qreal mMaxNormalizedWidth; //!< maximal normalized width, for all rows
-    QVector<QPair<int, int> > rowOffsets; //!< cached offsets for faster key lookups
+    QVector<QPair<qreal, qreal> > rowOffsets; //!< cached offsets for faster key lookups
     MImKey *shiftKey; //!< stores shift key, if available in this key area
     QTextLayout textLayout; //!< used to draw key labels onto key area
     bool textDirty; //!< dirty text cache flag
